refactor: @TestConfiguration -> @Configuration
main 실행 함수에서 @TestConfiguration을 읽지 못하는 문제로 rollback
