+++ src/test/java/team/silvertown/masil/MasilApplicationTests.java
@@ -7,10 +7,8 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
 
 @SpringBootTest
-@Import(value = {JwtTestConfig.class, OAuth2TestConfig.class})
 @DisplayNameGeneration(ReplaceUnderscores.class)
 class MasilApplicationTests {
